+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix loading of newer OpenNMT-py models
 * Promote FP16 to FP32 in model converter scripts
 * Improve INT8 performance on CPU and GPU
+* Enable INT8 and INT16 computation on all platforms with Intel MKL 2019.5 and above
 
 ## [v1.0.0](https://github.com/OpenNMT/CTranslate2/releases/tag/v1.0.0) (2019-09-23)